Move StartupFinalizerServiceComponent to Initializer
