+++ components/org.wso2.carbon.micro.integrator.initializer/pom.xml
@@ -127,6 +127,7 @@
                             org.wso2.carbon.micro.integrator.initializer.*;version="${project.version}"
                         </Export-Package>
                         <Import-Package>
+                            org.wso2.micro.integrator.core.*;version="${project.version}"
                             *;resolution:=optional
                         </Import-Package>
                         <DynamicImport-Package>*</DynamicImport-Package>
